Updates clamp() upon resizing of the viewport
This is important so that if the window size is changed, the viewport maintains its contraints.
e.g. If the viewport is at its max constraints and the window size increases, the viewport should clamp to those max contraints rather than allow the viewport to show more of the world.