Switch `yarn` for `npm`
The current setup requires `yarn` to be installed in
order to build this repo.

For the `pixi-v8` branch I'm pointing my app directly
at the branch via:

`"pixi-viewport": "git+https://github.com/vitch/pixi-viewport#pixi-v8"`

This works fine locally (where I installed `yarn`)
but fails on CI. So this commit attempts to remove the
dependency on `yarn`.

Note I also had to lock `electron` to a version compatible
with the `peerDependency` declared by `floss` (which
afaict is the only reason `electron` is here).